requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
chainlit>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
